refactor to operators and rules
