+++ __tests__/operators/chat-ops/review.test.ts
@@ -1,4 +1,4 @@
-import review from '../../../src/operations/chat-ops/review'
+import review from '../../../src/operators/chat-ops/review'
 import {Command, Commands} from "../../../src/command";
 import * as github from "@actions/github";
 import * as core from "@actions/core";
