+++ __tests__/operators/label.test.ts
@@ -1,4 +1,4 @@
-import label from '../../src/operations/label'
+import label from '../../src/operators/label'
 import {Command, Commands} from "../../src/command";
 import * as github from "@actions/github";
 import * as core from "@actions/core";
