+++ src/operators/capture.ts
